requirements.txt
streamlit
google-generativeai
opencv-python
python-dotenv
requests
numpy
ultralytics